// src/interfaces/IRespone.ts
export interface ILocation {
  name: string;
  region: string;
  country: string;
  lat: number;
  lon: number;
  tz_id: string;
  localtime_epoch: number;
  localtime: string;
}

// ===============================
// Weather Condition
// ===============================

export interface ICondition {
  text: string;
  icon: string;
  code: number;
}

// ===============================
// Air Quality
// ===============================

export interface IAirQuality {
  co: number;
  no2: number;
  o3: number;
  so2: number;
  pm2_5: number;
  pm10: number;

  // EPA index
  "us-epa-index": number;

  // UK DEFRA index
  "gb-defra-index": number;
}

// ===============================
// Current Weather
// ===============================

export interface ICurrentWeather {
  last_updated_epoch: number;
  last_updated: string;

  temp_c: number;
  temp_f: number;

  is_day: number;

  condition: ICondition;

  wind_mph: number;
  wind_kph: number;
  wind_degree: number;
  wind_dir: string;

  pressure_mb: number;
  pressure_in: number;

  precip_mm: number;
  precip_in: number;

  humidity: number;
  cloud: number;

  feelslike_c: number;
  feelslike_f: number;

  windchill_c: number;
  windchill_f: number;

  heatindex_c: number;
  heatindex_f: number;

  dewpoint_c: number;
  dewpoint_f: number;

  vis_km: number;
  vis_miles: number;

  uv: number;

  gust_mph: number;
  gust_kph: number;

  air_quality?: IAirQuality;
}
// ...
export interface IForecastDay {
  date: string;
  date_epoch: number;

  day: IDay;

  astro: IAstro;

  hour: IHour[];
}

// ===============================
// Forecast
// ===============================

export interface IForecast {
  forecastday: IForecastDay[];
}

// ===============================
// Weather Alerts
// ===============================

export interface IWeatherAlert {
  headline: string;
  msgtype: string;
  severity: string;
  urgency: string;
  areas: string;

  category: string;
  certainty: string;

  event: string;

  note: string;

  effective: string;
  expires: string;

  desc: string;
  instruction: string;

  web: string;
}

export interface IAlerts {
  alert: IWeatherAlert[];
}

// ===============================
// Main Weather Response
// ===============================

export interface IWeatherResponse {
  location: ILocation;

  current: ICurrentWeather;

  forecast: IForecast;

  alerts?: IAlerts;
}